refactor: moved specific component close to the router who uses it
